[bv2int] Add manual simplification
- After the changes, due to the unique template of reuftation phis,
we now get the template for phis that require translation:

Signed:
(x % 16 == 13) && (x <= 7) && (x >= -8)

Unsigned:
(x % 16 == 13) && (x <= 15) && (x >= 0)

Z3 fails to simplify the above expression. In case of signed BVs
we expect the simplification: x == -3
And in case of unsigned BVs we expect the simplification: x == 13.

This simplification is important as it simplifies the query expressions
passed to the fixedpoint solver behind the scenes. Simpler expressions
produce more general invariants which speed up out algorithm.